refactor: Update MoviesState interface and export moviesState
